Split tests_aggregate.py tests into multiple files
